Basic enemy AI complete
Set up enemy AI with 2 triggers--one that causes them to attack the player, and another that causes them to chase the player.

Enemies can either "leash" or not leash. Leashed enemies will return to their aggro area if they leave it.

Moved combat initiation to the weapon to get rid of OnTriggerEnter weirdness with the new AI trigger zones, and moves combat calculations to a manager class.